requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
ase==3.26.0
ase-db-backends==0.11.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bibtexparser==1.4.3
bleach==6.3.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
clusterscope==0.0.18
colorful==0.5.8
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
e3nn==0.5.8
executing==2.2.1
fairchem-core==2.13.0
fastapi==0.124.4
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.2.3
hydra-core==1.3.2
idna==3.11
importlib-metadata==8.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.7.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lightning==2.2.5
lightning-utilities==0.15.2
llvmlite==0.46.0
lmdb==1.7.3
markdown==3.10
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.1.4
monty==2025.3.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.0
notebook-shim==0.2.4
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt-einsum==3.4.0
opt-einsum-fx==0.1.4
orjson==3.11.5
packaging==24.2
palettable==3.3.3
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.3
py-spy==0.4.1
py3dmol==2.5.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyg-lib==0.5.0+pt26cu124
pygments==2.19.2
pymatgen==2025.10.7
pymysql==1.1.2
pyparsing==3.2.5
pyre-extensions==0.0.32
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytorch-lightning==2.6.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
ray==2.52.1
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.15
scikit-learn==1.8.0
scipy==1.16.3
send2trash==1.8.3
sentry-sdk==2.47.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.5.0
smmap==5.0.2
soupsieve==2.8
spglib==2.6.0
stack-data==0.6.3
starlette==0.50.0
submitit==1.5.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.6.0+cu124
torch-cluster==1.6.3+pt26cu124
torch-geometric==2.7.0
torch-scatter==2.1.2+pt26cu124
torch-sparse==0.6.18+pt26cu124
torch-spline-conv==1.2.2+pt26cu124
torchaudio==2.6.0+cu124
torchmetrics==1.8.2
torchtnt==0.2.4
torchvision==0.21.0+cu124
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
uncertainties==3.2.3
uri-template==1.3.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
wandb==0.23.1
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
